sensors/requirements.txt
paho-mqtt
setuptools
numpy